changed sprots.sql, and genericDownloader.php
https://app.asana.com/0/84113659891049/94393680990013
on sprots.sql, i changed the playerName row to varchar 100, according to the api documentation. On the downloader, i worked on getting it to grab the statistic values for the players, and divide them into the Statistic table, and from there divide them into the playerStatistic table. TODO, finish, player statistic import, and then do teamStatistic.